+++ src/traits/macros.rs
@@ -29,24 +29,40 @@ macro_rules! forward_methods {
 
         /// Forwards to [`BumpAllocator::scoped`].
         #[inline(always)]
-        pub fn scoped<R>(&mut $self, f: impl FnOnce(BumpScope<A, S>) -> R) -> R
-        where
-            S: BumpAllocatorSettings<GuaranteedAllocated = True>,
-        {
+        #[cfg(feature = "panic-on-alloc")]
+        pub fn scoped<R>(&mut $self, f: impl FnOnce(BumpScope<A, S>) -> R) -> R {
             BumpAllocator::scoped($access_mut, f)
+        }
+
+        /// Forwards to [`BumpAllocator::try_scoped`].
+        #[inline(always)]
+        pub fn try_scoped<R>(&mut $self, f: impl FnOnce(BumpScope<A, S>) -> R) -> Result<R, AllocError> {
+            BumpAllocator::try_scoped($access_mut, f)
         }
 
         /// Forwards to [`BumpAllocator::scoped_aligned`].
         #[inline(always)]
+        #[cfg(feature = "panic-on-alloc")]
         pub fn scoped_aligned<const NEW_MIN_ALIGN: usize, R>(
             &mut $self,
             f: impl FnOnce(BumpScope<A, S::WithMinimumAlignment<NEW_MIN_ALIGN>>) -> R,
         ) -> R
         where
             MinimumAlignment<NEW_MIN_ALIGN>: SupportedMinimumAlignment,
-            S: BumpAllocatorSettings<GuaranteedAllocated = True>,
         {
             BumpAllocator::scoped_aligned($access_mut, f)
+        }
+
+        /// Forwards to [`BumpAllocator::try_scoped_aligned`].
+        #[inline(always)]
+        pub fn try_scoped_aligned<const NEW_MIN_ALIGN: usize, R>(
+            &mut $self,
+            f: impl FnOnce(BumpScope<A, S::WithMinimumAlignment<NEW_MIN_ALIGN>>) -> R,
+        ) -> Result<R, AllocError>
+        where
+            MinimumAlignment<NEW_MIN_ALIGN>: SupportedMinimumAlignment,
+        {
+            BumpAllocator::try_scoped_aligned($access_mut, f)
         }
 
         /// Forwards to [`BumpAllocatorScope::aligned`].